WIP refactor workflows into subworkflows
There are already some redunancies between the different
"workflows" (salmon/alevin/star). This will be getting
worse as additional features downstream of the quantification
are getting implemented (e.g. conversion to AnnData, ambient
RNA filtering etc.)

This is the beginning of a series of commits that will
turn the workflows into subworkflows:
 - [ ] Streamline parameter validation and move to main workflow
 - [ ] Make the different alignment workflow subworkflows that are
       independent of `params` and get all input values via `take` instead
 - [ ] Make cellranger subworkflow available